Add files via upload
IoT is rapidly changing the healthcare industry and constantly expanding new healthcare
technologies. Monitoring the patient's health at home is a difficult task due to the busy
schedule of our daily work. Elderly patients in particular should be monitored regularly. Our
model establishes a patient health monitoring system using a web server and IoT applications
to measure patient health parameters such as heart rate (BPM) and blood oxygen level (SPo2)
along with human body temperature from anywhere. According to our model, we will
introduce the development of an IoT-based patient condition monitoring system. We will
have a pulse oximeter sensor to measure heart rate (BPM) and blood oxygen level (SpO2).
We will use a temperature sensor to measure body temperature. Therefore, it is necessary that
the patient be kept in a room with normal temperature and humidity so that he does not
experience any intolerance. To achieve this, we also need to monitor the temperature and
humidity in the room. So, we will use a humidity and temperature sensor. In this intelligent
patient monitoring system, patients' health is automatically measured through interconnected
sensor-based networks. Various sensors are used to collect the patient's biological parameters.
Meaningful biological information is then transmitted to the IoT server. Our system is so
smart that it can detect the vulnerable condition of the patient by processing the data collected
by the sensor. This system helps all doctors and their hospital faculties to virtually observe
assigned patients without physical visit. The patient's relatives also benefit through this
system.